🐛 fix so many migration errors???

+++ web-laravel-monetary-budget/app/Models/Budget.php
@@ -26,31 +26,40 @@ class Budget extends Model
         'category_ids'
     ];
 
-    public function getAccountIdsAttribute() {
+    public function getAccountIdsAttribute()
+    {
         return BudgetAccounts::query()->where("budget_id", $this->id)->pluck('account_id');
     }
 
-    public function setAccountIdsAttribute(array $accountIds) {
-        BudgetAccounts::query()->whereIn("account_id", $accountIds)->delete();
-        foreach ($accountIds as $accountId) {
-            BudgetAccounts::query()->create([
-                'budget_id' => $this->id,
-                'account_id' => $accountId
-            ]);
+    public function setAccountIdsAttribute(array $accountIds)
+    {
+        if (isset($this->id)) {
+            BudgetAccounts::query()->whereIn("account_id", $accountIds)->delete();
+            foreach ($accountIds as $accountId) {
+                BudgetAccounts::query()->create([
+                    'budget_id' => $this->id,
+                    'account_id' => $accountId
+                ]);
+            }
         }
+
     }
 
-    public function getCategoryIdsAttribute() {
+    public function getCategoryIdsAttribute()
+    {
         return BudgetCategories::query()->where("budget_id", $this->id)->pluck('category_id');
     }
 
-    public function setCategoryIdsAttribute(array $categoryIds) {
-        BudgetCategories::query()->whereIn("category_id", $categoryIds)->delete();
-        foreach ($categoryIds as $categoryId) {
-            BudgetCategories::query()->create([
-                'budget_id' => $this->id,
-                'category_id' => $categoryId
-            ]);
+    public function setCategoryIdsAttribute(array $categoryIds)
+    {
+        if (isset($this->id)) {
+            BudgetCategories::query()->whereIn("category_id", $categoryIds)->delete();
+            foreach ($categoryIds as $categoryId) {
+                BudgetCategories::query()->create([
+                    'budget_id' => $this->id,
+                    'category_id' => $categoryId
+                ]);
+            }
         }
     }
 }
