+++ web-laravel-monetary-debt/app/Models/Debt.php
@@ -27,17 +27,21 @@ class Debt extends Model
         'transaction_ids'
     ];
 
-    public function getTransactionIdsAttribute() {
+    public function getTransactionIdsAttribute()
+    {
         return DebtTransactions::query()->where('debt_id', $this->id)->pluck('transaction_id');
     }
 
-    public function setTransactionIdsAttribute(array $transactionIds) {
-        DebtTransactions::query()->whereIn("transaction_id", $transactionIds)->delete();
-        foreach ($transactionIds as $transactionId) {
-            DebtTransactions::query()->create([
-                'debt_id' => $this->id,
-                'transaction_id' => $transactionId
-            ]);
+    public function setTransactionIdsAttribute(array $transactionIds)
+    {
+        if (isset($this->id)) {
+            DebtTransactions::query()->whereIn("transaction_id", $transactionIds)->delete();
+            foreach ($transactionIds as $transactionId) {
+                DebtTransactions::query()->create([
+                    'debt_id' => $this->id,
+                    'transaction_id' => $transactionId
+                ]);
+            }
         }
     }
 }
